anda todo aveces la colision no la toma :/
